Separated signature verifier interface from specific implementation.
